refactor(models): split into package of many files
Take all models defined in benefits/core/models.py and split into files
to group models by domain:

* benefits/core/models/common.py: common fields/models, helper functions
* benefits/core/models/claims.py: ClaimsProvider model
* benefits/core/models/enrollment.py: EnrollmentFlow, EnrollmentEvent models
* benefits/core/models/transit.py: TransitProvider, TransitAgency models

Maintain existing imports via top-level benefits/core/models/__init__.py
